Refactor codebase into core and server subpackages
- Moved shared logic (config, exceptions, ignore_patterns, etc.) into `core/`.
- Created `server/` to host FastAPI modules, static files, and templates.
- Updated import paths for CLI and tests to reference `core.*` and `server.*`.
- Changed static references from `"/static"` to `"/server/static"` and templates from `"/templates"` to `"/server/templates"`.
- Adjusted references in `base.jinja`, `server/main.py`, `server/query_processor.py`, and routers to reflect the new server paths.
- Ensured Docker builds now correctly serve static and template files from `server/`.
