+++ src/core/exceptions.py
