+++ src/core/ignore_patterns.py
