+++ src/gitingest/__init__.py
@@ -1,8 +1,8 @@
 """ Gitingest: A package for ingesting data from Git repositories. """
 
-from gitingest.query_ingestion import run_ingest_query
-from gitingest.query_parser import parse_query
-from gitingest.repository_clone import clone_repo
-from gitingest.repository_ingest import ingest
+from core.query_ingestion import run_ingest_query
+from core.query_parser import parse_query
+from core.repository_clone import clone_repo
+from core.repository_ingest import ingest
 
 __all__ = ["run_ingest_query", "clone_repo", "parse_query", "ingest"]
