--- src/routers/__init__.py
@@ -1,7 +1,0 @@
-""" This module contains the routers for the FastAPI application. """
-
-from routers.download import router as download
-from routers.dynamic import router as dynamic
-from routers.index import router as index
-
-__all__ = ["download", "dynamic", "index"]
