+++ src/server/routers/download.py
@@ -3,7 +3,7 @@
 from fastapi import APIRouter, HTTPException
 from fastapi.responses import Response
 
-from config import TMP_BASE_PATH
+from core.config import TMP_BASE_PATH
 
 router = APIRouter()
