+++ src/server/routers/dynamic.py
@@ -1,49 +1,51 @@
-""" This module defines the FastAPI router for the home page of the application. """
+""" This module defines the dynamic router for handling dynamic path requests. """
 
 from fastapi import APIRouter, Form, Request
 from fastapi.responses import HTMLResponse
 from fastapi.templating import Jinja2Templates
 
-from config import EXAMPLE_REPOS
-from query_processor import process_query
-from server_utils import limiter
+from server.query_processor import process_query
+from server.server_utils import limiter
 
 router = APIRouter()
-templates = Jinja2Templates(directory="templates")
+templates = Jinja2Templates(directory="server/templates")
 
 
-@router.get("/", response_class=HTMLResponse)
-async def home(request: Request) -> HTMLResponse:
+@router.get("/{full_path:path}")
+async def catch_all(request: Request, full_path: str) -> HTMLResponse:
     """
-    Render the home page with example repositories and default parameters.
+    Render a page with a Git URL based on the provided path.
 
-    This endpoint serves the home page of the application, rendering the `index.jinja` template
-    and providing it with a list of example repositories and default file size values.
+    This endpoint catches all GET requests with a dynamic path, constructs a Git URL
+    using the `full_path` parameter, and renders the `git.jinja` template with that URL.
 
     Parameters
     ----------
     request : Request
         The incoming request object, which provides context for rendering the response.
+    full_path : str
+        The full path extracted from the URL, which is used to build the Git URL.
 
     Returns
     -------
     HTMLResponse
-        An HTML response containing the rendered home page template, with example repositories
-        and other default parameters such as file size.
+        An HTML response containing the rendered template, with the Git URL
+        and other default parameters such as loading state and file size.
     """
     return templates.TemplateResponse(
-        "index.jinja",
+        "git.jinja",
         {
             "request": request,
-            "examples": EXAMPLE_REPOS,
+            "repo_url": full_path,
+            "loading": True,
             "default_file_size": 243,
         },
     )
 
 
-@router.post("/", response_class=HTMLResponse)
+@router.post("/{full_path:path}", response_class=HTMLResponse)
 @limiter.limit("10/minute")
-async def index_post(
+async def process_catch_all(
     request: Request,
     input_text: str = Form(...),
     max_file_size: int = Form(...),
@@ -53,9 +55,8 @@ async def index_post(
     """
     Process the form submission with user input for query parameters.
 
-    This endpoint handles POST requests from the home page form. It processes the user-submitted
-    input (e.g., text, file size, pattern type) and invokes the `process_query` function to handle
-    the query logic, returning the result as an HTML response.
+    This endpoint handles POST requests, processes the input parameters (e.g., text, file size, pattern),
+    and calls the `process_query` function to handle the query logic, returning the result as an HTML response.
 
     Parameters
     ----------
@@ -73,7 +74,7 @@ async def index_post(
     Returns
     -------
     HTMLResponse
-        An HTML response containing the results of processing the form input and query logic,
+        An HTML response generated after processing the form input and query logic,
         which will be rendered and returned to the user.
     """
     return await process_query(
@@ -82,5 +83,5 @@ async def index_post(
         max_file_size,
         pattern_type,
         pattern,
-        is_index=True,
+        is_index=False,
     )
