+++ src/server/static/apple-touch-icon.png
