+++ src/server/static/js/utils.js
