+++ src/server/static/robots.txt
