+++ src/server/templates/components/footer.jinja
