+++ src/server/templates/components/navbar.jinja
