+++ tests/core/test_repository_clone.py
@@ -5,8 +5,8 @@ from unittest.mock import AsyncMock, patch
 
 import pytest
 
-from gitingest.exceptions import AsyncTimeoutError
-from gitingest.repository_clone import CloneConfig, _check_repo_exists, clone_repo
+from core.exceptions import AsyncTimeoutError
+from core.repository_clone import CloneConfig, _check_repo_exists, clone_repo
 
 
 @pytest.mark.asyncio
@@ -22,8 +22,8 @@ async def test_clone_repo_with_commit() -> None:
         branch="main",
     )
 
-    with patch("gitingest.repository_clone._check_repo_exists", return_value=True) as mock_check:
-        with patch("gitingest.repository_clone._run_git_command", new_callable=AsyncMock) as mock_exec:
+    with patch("core.repository_clone._check_repo_exists", return_value=True) as mock_check:
+        with patch("core.repository_clone._run_git_command", new_callable=AsyncMock) as mock_exec:
             mock_process = AsyncMock()
             mock_process.communicate.return_value = (b"output", b"error")
             mock_exec.return_value = mock_process
@@ -40,8 +40,8 @@ async def test_clone_repo_without_commit() -> None:
     """
     query = CloneConfig(url="https://github.com/user/repo", local_path="/tmp/repo", commit=None, branch="main")
 
-    with patch("gitingest.repository_clone._check_repo_exists", return_value=True) as mock_check:
-        with patch("gitingest.repository_clone._run_git_command", new_callable=AsyncMock) as mock_exec:
+    with patch("core.repository_clone._check_repo_exists", return_value=True) as mock_check:
+        with patch("core.repository_clone._run_git_command", new_callable=AsyncMock) as mock_exec:
             mock_process = AsyncMock()
             mock_process.communicate.return_value = (b"output", b"error")
             mock_exec.return_value = mock_process
@@ -63,7 +63,7 @@ async def test_clone_repo_nonexistent_repository() -> None:
         commit=None,
         branch="main",
     )
-    with patch("gitingest.repository_clone._check_repo_exists", return_value=False) as mock_check:
+    with patch("core.repository_clone._check_repo_exists", return_value=False) as mock_check:
         with pytest.raises(ValueError, match="Repository not found"):
             await clone_repo(clone_config)
             mock_check.assert_called_once_with(clone_config.url)
@@ -135,8 +135,8 @@ async def test_clone_repo_with_custom_branch() -> None:
         local_path="/tmp/repo",
         branch="feature-branch",
     )
-    with patch("gitingest.repository_clone._check_repo_exists", return_value=True):
-        with patch("gitingest.repository_clone._run_git_command", new_callable=AsyncMock) as mock_exec:
+    with patch("core.repository_clone._check_repo_exists", return_value=True):
+        with patch("core.repository_clone._run_git_command", new_callable=AsyncMock) as mock_exec:
             await clone_repo(clone_config)
             mock_exec.assert_called_once_with(
                 "git",
@@ -160,8 +160,8 @@ async def test_git_command_failure() -> None:
         url="https://github.com/user/repo",
         local_path="/tmp/repo",
     )
-    with patch("gitingest.repository_clone._check_repo_exists", return_value=True):
-        with patch("gitingest.repository_clone._run_git_command", side_effect=RuntimeError("Git command failed")):
+    with patch("core.repository_clone._check_repo_exists", return_value=True):
+        with patch("core.repository_clone._run_git_command", side_effect=RuntimeError("Git command failed")):
             with pytest.raises(RuntimeError, match="Git command failed"):
                 await clone_repo(clone_config)
 
@@ -176,8 +176,8 @@ async def test_clone_repo_default_shallow_clone() -> None:
         url="https://github.com/user/repo",
         local_path="/tmp/repo",
     )
-    with patch("gitingest.repository_clone._check_repo_exists", return_value=True):
-        with patch("gitingest.repository_clone._run_git_command", new_callable=AsyncMock) as mock_exec:
+    with patch("core.repository_clone._check_repo_exists", return_value=True):
+        with patch("core.repository_clone._run_git_command", new_callable=AsyncMock) as mock_exec:
             await clone_repo(clone_config)
             mock_exec.assert_called_once_with(
                 "git", "clone", "--depth=1", "--single-branch", clone_config.url, clone_config.local_path
@@ -195,8 +195,8 @@ async def test_clone_repo_commit_without_branch() -> None:
         local_path="/tmp/repo",
         commit="a" * 40,  # Simulating a valid commit hash
     )
-    with patch("gitingest.repository_clone._check_repo_exists", return_value=True):
-        with patch("gitingest.repository_clone._run_git_command", new_callable=AsyncMock) as mock_exec:
+    with patch("core.repository_clone._check_repo_exists", return_value=True):
+        with patch("core.repository_clone._run_git_command", new_callable=AsyncMock) as mock_exec:
             await clone_repo(clone_config)
             assert mock_exec.call_count == 2  # Clone and checkout calls
             mock_exec.assert_any_call("git", "clone", "--single-branch", clone_config.url, clone_config.local_path)
@@ -245,8 +245,8 @@ async def test_clone_repo_with_timeout() -> None:
     """
     clone_config = CloneConfig(url="https://github.com/user/repo", local_path="/tmp/repo")
 
-    with patch("gitingest.repository_clone._check_repo_exists", return_value=True):
-        with patch("gitingest.repository_clone._run_git_command", new_callable=AsyncMock) as mock_exec:
+    with patch("core.repository_clone._check_repo_exists", return_value=True):
+        with patch("core.repository_clone._run_git_command", new_callable=AsyncMock) as mock_exec:
             mock_exec.side_effect = asyncio.TimeoutError
             with pytest.raises(AsyncTimeoutError, match="Operation timed out after"):
                 await clone_repo(clone_config)
